+++ src/BDSCOD/Utility.hs
@@ -2,34 +2,36 @@ module BDSCOD.Utility where
 
 import qualified Data.Vector as V
 import qualified Epidemic as EpiSim
-import Epidemic.Types
+import Epidemic.Types.Parameter
+import Epidemic.Types.Events
+import Epidemic.Types.Population
 import BDSCOD.Types
 -- import BDSCOD.Llhd --
 
 -- | Convert simulation events to observation events
-eventsAsObservations :: [EpiSim.Event] -> [Observation]
+eventsAsObservations :: [EpidemicEvent] -> [Observation]
 eventsAsObservations epiSimEvents =
-  drop 1 . map fst $ scanl processEvent' ((0, Birth), 0) epiSimEvents
+  drop 1 . map fst $ scanl processEvent' ((0, OBirth), 0) epiSimEvents
 
-processEvent' :: (Observation, Time) -> EpiSim.Event -> (Observation, Time)
+processEvent' :: (Observation, Time) -> EpidemicEvent -> (Observation, Time)
 processEvent' (_, currTime) epiSimEvent =
   case epiSimEvent of
-    (EpiSim.InfectionEvent absTime _ _) ->
-      ((absTime - currTime, Birth), absTime)
-    (EpiSim.SamplingEvent absTime _) -> ((absTime - currTime, Sample), absTime)
-    (EpiSim.CatastropheEvent absTime (EpiSim.People persons)) -> ((absTime - currTime, Catastrophe . fromIntegral $ V.length persons), absTime)
-    (EpiSim.OccurrenceEvent absTime _) ->
-      ((absTime - currTime, Occurrence), absTime)
-    (EpiSim.DisasterEvent absTime (EpiSim.People persons)) -> ((absTime - currTime, Disaster . fromIntegral $ V.length persons), absTime)
-    (EpiSim.RemovalEvent _ _) -> error "A removal event has been passed to processEvent', this should never happen!"
+    (Infection absTime _ _) ->
+      ((absTime - currTime, OBirth), absTime)
+    (Sampling absTime _) -> ((absTime - currTime, OSample), absTime)
+    (Catastrophe absTime (People persons)) -> ((absTime - currTime, OCatastrophe . fromIntegral $ V.length persons), absTime)
+    (Occurrence absTime _) ->
+      ((absTime - currTime, OOccurrence), absTime)
+    (Disaster absTime (People persons)) -> ((absTime - currTime, ODisaster . fromIntegral $ V.length persons), absTime)
+    (Removal _ _) -> error "A removal event has been passed to processEvent', this should never happen!"
 
 -- | Predicate for the observation referring to a birth.
 isBirth :: Observation -> Bool
-isBirth (_,e) = e == Birth
+isBirth (_,e) = e == OBirth
 
 -- | Predicate for the observation referring to a sampling.
 isSample :: Observation -> Bool
-isSample (_,e) = e == Sample
+isSample (_,e) = e == OSample
 
 nbFromMAndV :: (Double, Double) -> NegativeBinomial
 nbFromMAndV (0, 0) = Zero
@@ -52,15 +54,18 @@ nbPGF nb z = case nb of
   Zero -> 1
   (NegBinom r p) -> ((1 - p) / (1 - p * z)) ** r
 
+nbPGF' :: NegativeBinomial -> Double -> Double
 nbPGF' nb z = case nb of
   Zero -> 0
   (NegBinom r p) -> (r * p / (1 - p)) * nbPGF (NegBinom (r+1) p) z
 
+nbPGF'' :: NegativeBinomial -> Double -> Double
 nbPGF'' nb z = case nb of
   Zero -> 0
   (NegBinom r p) -> (r * (r + 1) * (p / (1 - p)) ** 2.0) *
                       nbPGF (NegBinom (r+2) p) z
 
+nbPGFdash :: Double -> NegativeBinomial -> Double -> Double
 nbPGFdash j nb z =
   case nb of
     Zero -> 0
@@ -68,5 +73,6 @@ nbPGFdash j nb z =
       pochhammer r j * (p / (1 - p)) ** j *
         nbPGF (NegBinom (r + j) p) z
 
+pochhammer :: (Eq p, Num p) => p -> p -> p
 pochhammer _ 0 = 1
 pochhammer a i = (a + i - 1) * pochhammer a (i - 1)